+++ setup.py
@@ -35,18 +35,11 @@ setup(name='speechtools',
       author='Montreal Corpus Tools',
       author_email='michael.e.mcauliffe@gmail.com',
       packages=['speechtools',
-                'speechtools.acoustics',
-                'speechtools.graph',
-                'speechtools.graph.attributes',
-                'speechtools.gui',
-                'speechtools.gui.plot',
-                'speechtools.gui.plot.widgets',
-                'speechtools.gui.widgets',
-                'speechtools.gui.widgets.query',
-                'speechtools.io',
-                'speechtools.sql'],
+                'speechtools.plot',
+                'speechtools.plot.widgets',
+                'speechtools.widgets',
+                'speechtools.widgets.query'],
       install_requires=[
-          'acousticsim',
           #'polyglotdb',
           #'vispy',
           'librosa',
