+++ speechtools/__init__.py
@@ -4,10 +4,3 @@ __ver_patch__ = 0
 __ver_tuple__ = (__ver_major__, __ver_minor__, __ver_patch__)
 __version__ = "%d.%d.%d" % __ver_tuple__
 
-__all__ = ['graph', 'sql', 'acoustics']
-
-import speechtools.graph as graph
-
-import speechtools.sql as sql
-
-import speechtools.acoustics as acoustics
