--- speechtools/acoustics/__init__.py
@@ -1,3 +1,0 @@
-
-
-from .analysis import acoustic_analysis
