--- speechtools/exceptions.py
@@ -1,15 +1,0 @@
-import os
-import sys
-import traceback
-
-from polyglotdb.exceptions import *
-
-## Base exceptions
-
-class SCTError(Exception):
-    pass
-
-## Acoustic exceptions
-
-class NoSoundFileError(SCTError):
-    pass
