--- speechtools/graph/attributes/__init__.py
@@ -1,2 +1,0 @@
-
-from .pause import PauseAnnotation, PausePathAnnotation
