--- speechtools/gui/__init__.py
@@ -1,1 +1,0 @@
-
