+++ speechtools/helper.py
