+++ speechtools/plot/cameras.py
