+++ speechtools/plot/widgets/annotation.py
