+++ speechtools/plot/widgets/base.py
