+++ speechtools/plot/widgets/spectrogram.py
