+++ speechtools/widgets/base.py
