+++ speechtools/widgets/details.py
