--- tests/data/textgrids/phone_word_no_silence.TextGrid
@@ -1,44 +1,0 @@
-File type = "ooTextFile"
-Object class = "TextGrid"
-
-xmin = 0 
-xmax = 1 
-tiers? <exists> 
-size = 2 
-item []: 
-    item [1]:
-        class = "IntervalTier" 
-        name = "phone" 
-        xmin = 0 
-        xmax = 1 
-        intervals: size = 4 
-        intervals [1]:
-            xmin = 0 
-            xmax = 0.25 
-            text = "a" 
-        intervals [2]:
-            xmin = 0.25 
-            xmax = 0.5 
-            text = "a" 
-        intervals [3]:
-            xmin = 0.5 
-            xmax = 0.75 
-            text = "b" 
-        intervals [4]:
-            xmin = 0.75 
-            xmax = 1 
-            text = "b" 
-    item [2]:
-        class = "IntervalTier" 
-        name = "word" 
-        xmin = 0 
-        xmax = 1 
-        intervals: size = 2 
-        intervals [1]:
-            xmin = 0 
-            xmax = 0.5 
-            text = "a" 
-        intervals [2]:
-            xmin = 0.5 
-            xmax = 1 
-            text = "b"
