+++ tests/test_gui_widgets.py
@@ -1,16 +1,15 @@
 
 import pytest
 
-from speechtools.gui.widgets.annotation import SubannotationDialog, NoteDialog
-from speechtools.gui.widgets.audio import AudioOutput, Generator
-from speechtools.gui.widgets.base import CollapsibleWidgetPair, DataListWidget, DetailedMessageBox
-from speechtools.gui.widgets.connection import ConnectWidget, CorporaList
-from speechtools.gui.widgets.details import DetailsWidget
-from speechtools.gui.widgets.help import HelpWidget
-from speechtools.gui.widgets.main import DiscourseWidget, ViewWidget
-from speechtools.gui.widgets.query import QueryWidget, QueryForm, QueryResults
-from speechtools.gui.widgets.selectable_audio import SelectableAudioWidget
-from speechtools.gui.widgets.structure import HierarchyWidget
+from speechtools.widgets.annotation import SubannotationDialog, NoteDialog
+from speechtools.widgets.audio import AudioOutput, Generator
+from speechtools.widgets.base import CollapsibleWidgetPair, DataListWidget, DetailedMessageBox
+from speechtools.widgets.connection import ConnectWidget, CorporaList
+from speechtools.widgets.details import DetailsWidget
+from speechtools.widgets.help import HelpWidget
+from speechtools.widgets.main import DiscourseWidget, ViewWidget
+from speechtools.widgets.selectable_audio import SelectableAudioWidget
+from speechtools.widgets.structure import HierarchyWidget
 
 def test_subannotation_dialog(qtbot):
     w = SubannotationDialog()
@@ -47,19 +46,6 @@ def test_discourse_widget(qtbot):
 
 def test_view_widget(qtbot):
     w = ViewWidget()
-    qtbot.addWidget(w)
-
-def test_query_widget(qtbot):
-    w = QueryWidget()
-    qtbot.addWidget(w)
-
-def test_query_form(qtbot):
-    w = QueryForm()
-    qtbot.addWidget(w)
-
-@pytest.mark.xfail
-def test_query_results(qtbot):
-    w = QueryResults()
     qtbot.addWidget(w)
 
 def test_selectable(qtbot):
